+++ pkg/tfbridge/schema.go
@@ -461,13 +461,30 @@ const metaKey = "__meta"
 // MakeTerraformResult expands a Terraform state into an expanded Pulumi resource property map.  This respects
 // the property maps so that results end up with their correct Pulumi names when shipping back to the engine.
 func MakeTerraformResult(state *terraform.InstanceState,
-	tfs map[string]*schema.Schema, ps map[string]*SchemaInfo) resource.PropertyMap {
+	tfs map[string]*schema.Schema, ps map[string]*SchemaInfo) (resource.PropertyMap, error) {
 	var outs map[string]interface{}
 	if state != nil {
 		outs = make(map[string]interface{})
 		attrs := state.Attributes
+
+		reader := &schema.MapFieldReader{
+			Schema: tfs,
+			Map:    schema.BasicMapReader(attrs),
+		}
 		for _, key := range flatmap.Map(attrs).Keys() {
-			outs[key] = flatmap.Expand(attrs, key)
+			res, err := reader.ReadField([]string{key})
+			if err != nil {
+				return nil, err
+			}
+			if res.Value != nil {
+				outs[key] = res.Value
+			}
+		}
+
+		// Populate the "id" property if it is not set. Most schemas do not include this property, and leaving it out
+		// can cause unnecessary diffs when refreshing/updating resources after a provider upgrade.
+		if _, ok := outs["id"]; !ok {
+			outs["id"] = attrs["id"]
 		}
 	}
 	outMap := MakeTerraformOutputs(outs, tfs, ps, nil, false)
@@ -479,7 +496,7 @@ func MakeTerraformResult(state *terraform.InstanceState,
 		outMap[metaKey] = resource.NewStringProperty(string(metaJSON))
 	}
 
-	return outMap
+	return outMap, nil
 }
 
 // MakeTerraformOutputs takes an expanded Terraform property map and returns a Pulumi equivalent.  This respects
@@ -522,6 +539,11 @@ func MakeTerraformOutput(v interface{},
 
 	if v == nil {
 		return resource.NewNullProperty()
+	}
+
+	// Marshal sets as their list value.
+	if set, isset := v.(*schema.Set); isset {
+		v = set.List()
 	}
 
 	// We use reflection instead of a type switch so that we can support mapping values whose underlying type is
@@ -615,7 +637,7 @@ func MakeTerraformOutput(v interface{},
 		obj := MakeTerraformOutputs(outs, tfflds, psflds, assets, rawNames || useRawNames(tfs))
 		return resource.NewObjectProperty(obj)
 	default:
-		contract.Failf("Unexpected TF output property value: %v", v)
+		contract.Failf("Unexpected TF output property value: %#v", v)
 		return resource.NewNullProperty()
 	}
 }