+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This CHANGELOG details important changes made in each version of the
 
 ## v0.18.3 (Unreleased)
 
+- Fixed a bug that caused unnecessary changes if the first operation after upgrading a bridged provider was a `pulumi refresh`.
+
 ### Improvements
 
 - Automatically generate `isInstance` type guards for implementations of `Resource`.